requirements.txt
arrow==0.12.1
backcall==0.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
decorator==4.3.0
idna==2.6
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
misaka==2.1.0
oauthlib==2.1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-decouple==3.1
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
tapioca-github==0.3
tapioca-wrapper==1.4.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xmltodict==0.11.0
